- PrePersist is seeming to be the wrong callback for TT - Validation should happen on raw data passed to the entity instead of processed values
